Google Search Algorithm
PageRank is a link analysis algorithm and it assigns a numerical
weighting to each element of a hyperlinked set of documents, such as the
World Wide Web, with the purpose of "measuring" its relative importance
within the set. The algorithm may be applied to any collection of
entities with reciprocal quotations and references. The numerical weight
that it assigns to any given element E is referred to as the PageRank of
E and denoted by PR(E). Other factors like Author Rank can contribute to
the importance of an entity.

A PageRank results from a mathematical algorithm based on the webgraph,
created by all World Wide Web pages as nodes and hyperlinks as edges,
taking into consideration authority hubs such as cnn.com or usa.gov. The
rank value indicates an importance of a particular page. A hyperlink to
a page counts as a vote of support. The PageRank of a page is defined
recursively and depends on the number and PageRank metric of all pages
that link to it ("incoming links"). A page that is linked to by many
pages with high PageRank receives a high rank itself.

Numerous academic papers concerning PageRank have been published since
Page and Brin's original paper.[5] In practice, the PageRank concept may
be vulnerable to manipulation. Research has been conducted into
identifying falsely influenced PageRank rankings. The goal is to find an
effective means of ignoring links from documents with falsely influenced
PageRank.[6]

Other link-based ranking algorithms for Web pages include the HITS
algorithm invented by Jon Kleinberg (used by Teoma and now
Ask.com),[citation needed] the IBM CLEVER project, the TrustRank
algorithm and the hummingbird algorithm.